fndataops/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
duckdb==0.9.2
pandas==2.1.4
numpy==1.24.4

# Data processing
dbt-core==1.7.0
dbt-postgres==1.7.0
great-expectations==0.18.0

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0

# Data generation and testing
faker==20.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Security and monitoring
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data quality and validation
pandera==0.17.2
cerberus==1.3.5

# Visualization and analysis
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jupyter==1.0.0
notebook==7.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
ruff==0.1.7
mypy==1.7.1

# Security scanning
bandit==1.7.5
safety==2.3.5

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0

# API documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0